// generated/api.ts
export interface components {
  schemas: {
    /** @description Info about checking inn */
    checkInnRequest: {
      /** @example 7725296566 */
      inn?: string;
      /** @example OOO TEST-Company */
      legal_name?: string;
      /** @example 109544, Москва г */
      legal_address?: string;
      /** @example Антонов Андрей Андреевич */
      legal_contact_name?: Record<string, never>;
      /** @example 7725296566 */
      kpp?: string;
    };
    /** @description Response after login */
    login: {
      /** @example WJJZFbwN6H2uKivqMr5Mu5MAIdkdaPbPcjWE7lxJQTDtONT8ZVqXLFrpTgUW */
      token?: string;
      /** @example true */
      is_verified?: boolean;
    };
    /** @description User info */
    userinfo: {
      /** @example 00e46820-b54f-49c0-bcbb-43909168bf11 */
      id?: string;
      /** @example some-user@gmail.com */
      email?: string;
      /** @example manager */
      role?: string;
      /** @example true */
      is_verified?: boolean;
      /** @example true */
      privacy_policy_accepted?: boolean;
      /** @example true */
      user_agreement_accepted?: boolean;
      /** @example true */
      is_active?: boolean;
    };
    /** @description User account in service */
    userAccount: {
      /** @example 00e46820-b54f-49c0-bcbb-43909168bf11 */
      id?: string;
      /** @example 00e46820-b54f-49c0-bcbb-43909168bf11 */
      user_id?: string;
      /** @example mytarget */
      service?: string;
      /** @example pending */
      status?: string;
      /** @example 2020-12-30T07:57:00.000000Z */
      created_at?: string;
      /** @example 2020-12-30T07:57:00.000000Z */
      updated_at?: string;
    };
    /** @description Info about checking account */
    checkAccount: {
      /** @example active */
      status?: string;
      /** @example Ошибка от сервиса */
      error?: string;
    };
    /** @description Info about business account */
    businessAccount: {
      /** @example c575deef-361a-4873-a579-65582e0884c5 */
      id?: string;
      /** @example ae912748-7e12-46ba-99c2-fa7e3701ab2f */
      user_id?: string;
      /** @example true */
      is_active?: boolean;
      /**
       * @example {
       *   "vkontakte": "default",
       *   "mytarget": "default"
       * }
       */
      agency_account_type?: Record<string, never>;
      /** @example Петров */
      first_name?: string;
      /** @example Иван */
      last_name?: string;
      /** @example Сергеевич */
      patronymic?: string;
      /** @example serg123@gmail.com */
      email?: string;
      /** @example +3729987548 */
      phone?: string;
      /** @example mail */
      receiving_checks_method?: string;
      /**
       * Format: float
       * @example 5000.0000
       */
      balance?: number;
      /**
       * Format: float
       * @example 30.0000
       */
      cashback_balance?: number;
      /** @example true */
      is_legal?: boolean;
      /** @example true */
      offer_agreement_accepted?: boolean;
      /** @example 7710140679 */
      inn?: string;
      /** @example 7710140679 */
      kpp?: string;
      /** @example 044525974 */
      bic?: string;
      /** @example 40817810100000670946 */
      iban?: string;
      /** @example WFBIUS6S */
      swift_code?: string;
      /** @example АО Тинькофф Банк */
      bank_name?: string;
      /** @example Москва, 123060, 1-й Волоколамский проезд */
      bank_address?: string;
      /** @example Russia */
      legal_country?: string;
      /** @example ИП Иванов */
      legal_name?: string;
      /** @example 143900, Россия, г. Балашиха, ул. Некрасова 15-20 */
      legal_address?: string;
      /** @example Иванов П.П. */
      legal_contact_name?: string;
      /** @example mail */
      legal_receiving_docs_method?: string;
      /**
       * @description Индекс почтового адреса
       * @example 222827
       */
      legal_receiving_mail_index?: string;
      /**
       * @description почта получателя
       * @example test@gmail.com
       */
      legal_receiving_mail_address?: string;
      /**
       * @description получатель
       * @example Иванов Иван Иванович
       */
      legal_receiving_mail_recipient?: string;
      /**
       * @description физический адрес
       * @example ул. Тестовая, д. 42, кв. 37
       */
      legal_receiving_physical_address?: string;
      /** @example 2020-12-30T07:57:00.000000Z */
      created_at?: string;
      /** @example 2020-12-30T07:57:00.000000Z */
      updated_at?: string;
      /**
       * @description Наценка на курс
       * @example 5
       */
      course_percent?: Record<string, never>;
      /**
       * @example {
       *   "cilent_id": "92ca65cd-dc14-4851-a1d5-26696e075f9c",
       *   "cilent_secret": "3jKlp6Q9pADe3AevgVfgi3QwNmFPM8P4Kd9Xo674"
       * }
       */
      oaut2_data?: Record<string, never>;
      /** @example 745454112 */
      tracker_manager_id?: string;
      /** @example true */
      company_verify?: boolean;
      /**
       * @example {
       *   "deposit": [
       *     {
       *       "max": null,
       *       "min": 0,
       *       "percent": 0
       *     }
       *   ],
       *   "advertising": {
       *     "mytarget": [
       *       {
       *         "max": null,
       *         "min": 0,
       *         "percent": 0
       *       }
       *     ],
       *     "vkontakte": [
       *       {
       *         "max": null,
       *         "min": 0,
       *         "percent": 0
       *       }
       *     ]
       *   },
       *   "commission": {
       *     "mytarget": [
       *       {
       *         "max": null,
       *         "min": 0,
       *         "percent": 0
       *       }
       *     ],
       *     "facebook": [
       *       {
       *         "max": null,
       *         "min": 0,
       *         "percent": 0
       *       }
       *     ],
       *     "vkontakte": [
       *       {
       *         "max": null,
       *         "min": 0,
       *         "percent": 0
       *       }
       *     ]
       *   }
       * }
       */
      cashback_conditions?: Record<string, never>;
      /**
       * @example {
       *   "for_advertising": "1000",
       *   "for_deposits": "500",
       *   "total_cashback": "1500"
       * }
       */
      current_month_cashback?: Record<string, never>;
      /**
       * @example {
       *   "vkontakte": false,
       *   "mytarget": true
       * }
       */
      visible_services?: Record<string, never>;
    };
    FieldsValidationError: {
      /** @example The given data was invalid. */
      message?: string;
      /**
       * @example {
       *   "field": [
       *     [
       *       "field error message"
       *     ]
       *   ]
       * }
       */
      errors?: Record<string, never>;
    };
    OrderPayLink: {
      /** @example POST */
      method?: string;
      /** @example https://capitalist.net/merchant/pay */
      url?: string;
      /**
       * @example {
       *   "merchantid": "12345",
       *   "number": "df11349d-1b44-4db4-87b6-7e0fbfbed0fd",
       *   "amount": "123",
       *   "currency": "USD"
       * }
       */
      data?: Record<string, never>;
    };
    Payment: {
      /** @example true */
      payed?: boolean;
    };
    EloquentPaginationLinks: {
      /** @example http://lk.local/api/v1/entity?page=1 */
      first?: string;
      /** @example http://lk.local/api/v1/entity?page=99 */
      last?: string;
      /** @example http://lk.local/api/v1/entity?page=1 */
      prev?: string;
      /** @example http://lk.local/api/v1/entity?page=2 */
      next?: string;
    };
    EloquenPaginationMetaLink: {
      /** @example http://lk.local/api/v1/entity?page=1 */
      url?: string;
      /** @example &laquo; Назад */
      label?: string;
      /** @example false */
      active?: boolean;
    };
    EloquentPaginationMeta: {
      /** @example 1 */
      current_page?: number;
      /** @example 1 */
      from?: number;
      /** @example 100 */
      last_page?: number;
      links?: components['schemas']['EloquenPaginationMetaLink'][];
      /** @example http://lk.local/api/v1/entity */
      path?: string;
      /** @example 10 */
      per_page?: number;
      /** @example 4 */
      to?: number;
      /** @example 4 */
      total?: number;
    };
    OrderStatus: {
      /** @example c575deef-361a-4873-a579-65582e0884c5 */
      id?: string;
      /** @example inprogress */
      code?: string;
      /** @example In Progress */
      title?: string;
      /** @example true */
      default?: boolean;
      /** @example false */
      finally?: boolean;
    };
    Currency: {
      /** @example USD */
      code?: string;
      /** @example Dollar $ */
      title?: string;
    };
    /** @description Payment gateway */
    Gateway: {
      /** @example c575deef-361a-4873-a579-65582e0884c5 */
      id?: string;
      /** @example capitalist */
      code?: string;
      /** @example Capitalist */
      title?: string;
      /** @example Pay via Capitalist */
      description?: string;
      /** @example 99 */
      sort?: number;
      /** @example /images/logos/cap.png */
      logo?: string;
      currencies?: components['schemas']['Currency'][];
    };
    /** @description Info about orders */
    Order: {
      /** @example c575deef-361a-4873-a579-65582e0884c5 */
      id?: string;
      /** @example ae912748-7e12-46ba-99c2-fa7e3701ab2f */
      business_account_id?: string;
      /** @example ae912748-7e12-46ba-99c2-fa7e3701ab2f */
      gateway_id?: string;
      gateway?: components['schemas']['Gateway'];
      /** @example ae912748-7e12-46ba-99c2-fa7e3701ab2f */
      order_status_id?: string;
      order_status?: components['schemas']['OrderStatus'];
      /** @example USD */
      currency?: string;
      /**
       * Format: float
       * @example 100.00
       */
      sum?: number;
      /** @example 2020-12-30T07:57:00.000000Z */
      created_at?: string;
      /** @example 2020-12-30T07:57:00.000000Z */
      updated_at?: string;
    };
    Transaction: {
      /** @example c575deef-361a-4873-a579-65582e0884c5 */
      id?: string;
      /** @example transfer */
      type?: string;
      /** @example Зачисление */
      name?: string;
      /** @example Пополнение рекл.аккаунта slaftestlog-5@gmail.com на сумму 1 RUB. */
      desc?: string;
      /**
       * Format: float
       * @example 100.00
       */
      amountTotal?: number;
      /** @example 2020-12-30T07:57:00.000000Z */
      date?: string;
    };
    Balance: {
      /**
       * Format: float
       * @example 100.00
       */
      balance?: number;
    };
    AdditionalManager: {
      /** @example c575deef-361a-4873-a579-65582e0884c5 */
      id?: string;
      /** @example mytarget_additional2@manager.com */
      manager_name?: string;
      /** @example 9338c95c-c6f7-460f-9b6f-05e03ebdde18 */
      business_account_id?: string;
      /** @example mytarget */
      service?: string;
      /** @example active */
      status?: string;
      /** @example Some error message after create manager */
      decline_message?: string;
      /** @example 2020-12-30T07:57:00.000000Z */
      created_at?: string;
      /** @example 2020-12-30T07:57:00.000000Z */
      updated_at?: string;
    };
    /** @description Finance Document entity */
    FinanceDocument: {
      /** @example 00e46820-b54f-49c0-bcbb-43909168bf11 */
      id?: string;
      /** @example 1 */
      number?: number;
      /** @example invoice */
      file_type?: string;
      /** @example https://site.come/get-temp-link/00e46820-b54f-49c0-bcbb-43909168bf11 */
      file_link?: string;
      /** @example invoice.pdf */
      file_name?: string;
      /**
       * Format: date
       * @example 2022-03-03
       */
      store_at?: string;
    };
// ...
    CurrencyRateResource: {
      /** @example USD */
      from?: string;
      /** @example RUB */
      to?: string;
      /**
       * Format: float
       * @example 76.125
       */
      value?: number;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}